rename to category.model.ts
