+++ src/app/tasks/components/single-category/single-category.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit, OnDestroy } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { Subject, takeUntil } from 'rxjs';
 import { categories } from '../../data/categories';
-import { Category } from '../../models/category';
+import { Category } from '../../models/category.model';
 
 @Component({
   selector: 'app-single-category',
